feat: Resolve image and container name conflict
- Unify naming convention between compose.yml and Makefile
- Implement dynamic name generation in Makefile
- Use environment variables for consistent naming across files
